Refactor relative data into own files, depluralize
