+++ src/apps/students/api/urls.py
@@ -1,5 +1,5 @@
-from django.urls import path
 from rest_framework.routers import DefaultRouter
+
 from apps.students.api.views import (
     StudentViewSet,
     RelativeViewSet,
