+++ src/apps/students/api/views/__init__.py
@@ -1,1 +1,2 @@
-from .students_view_set import StudentViewSet, RelativeViewSet
+from .student_view_set import StudentViewSet
+from .relative_view_set import RelativeViewSet
