+++ src/apps/students/api/views/student_view_set.py
@@ -9,10 +9,6 @@ from apps.students.api.serializers import (
     StudentViewSerializer,
     StudentCreationSerializer,
 )
-from apps.students.api.serializers import (
-    RelativeViewSerializer,
-    RelativeCreationSerializer,
-)
 
 
 class StudentViewSet(GenericModelViewSet):
@@ -64,18 +60,3 @@ class StudentViewSet(GenericModelViewSet):
         return Response(
             {"detail": "No student found"}, status=status.HTTP_400_BAD_REQUEST
         )
-
-
-class RelativeViewSet(GenericModelViewSet):
-    """
-    Generic ViewSet for Relative Model
-        - GET: list all relatives
-        - POST: create a relative
-        - GET(id): get a relative by id
-        - PUT(id): update a relative by id
-        - DELETE(id): delete a relative by id
-    """
-
-    serializer_class = RelativeViewSerializer
-    create_serializer_class = RelativeCreationSerializer
-    update_serializer_class = RelativeCreationSerializer
